fix: move theme types to its proper folder
